+++ config/settings/test.py
@@ -51,6 +51,7 @@ CACHES = {
 # TESTING
 # ------------------------------------------------------------------------------
 TEST_RUNNER = 'django.test.runner.DiscoverRunner'
+SILENCED_SYSTEM_CHECKS = ['axes.W003']  # Silence missing axes backend warning
 
 # PASSWORD HASHING
 # ------------------------------------------------------------------------------
@@ -78,10 +79,14 @@ REST_FRAMEWORK['PAGE_SIZE'] = 1
 
 # AUTHENTICATION CONFIGURATION
 # ------------------------------------------------------------------------------
+
+# NOTE: Hardcoded due to issue #1767
+# NOTE: Axes doesn't work with UI tests and cookie-based login, see #1810
+#       Override with AUTHENTICATION_BACKENDS_AXES when testing Axes features
 AUTHENTICATION_BACKENDS = [
     'rules.permissions.ObjectPermissionBackend',
     'django.contrib.auth.backends.ModelBackend',
-]  # NOTE: Hardcoded due to issue #1767
+]
 
 # LDAP configuration
 # ------------------------------------------------------------------------------
@@ -93,12 +98,29 @@ ENABLE_LDAP = False
 
 ENABLE_OIDC = False
 
+# Django-Axes
+# ------------------------------------------------------------------------------
+
+AXES_ENABLED = False  # Enable by override when testing
+AXES_FAILURE_LIMIT = 3
+AXES_LOCK_OUT_AT_FAILURE = True
+AXES_COOLOFF_TIME = None
+AXES_LOCKOUT_PARAMETERS = ['username']
+AXES_ONLY_ADMIN_SITE = False
+AXES_CLIENT_IP_CALLABLE = lambda x: None  # noqa: E731
+
 # Logging
 # ------------------------------------------------------------------------------
 
 LOGGING_LEVEL = env.str('LOGGING_LEVEL', 'CRITICAL')
 LOGGING = set_logging(LOGGING_LEVEL)
+LOGGING['loggers']['axes'] = {
+    'level': LOGGING_LEVEL,
+    'handlers': ['console'],
+    'propagate': False,
+}  # Disable redundant axes logging in tests
 LOGGING_DISABLE_CMD_OUTPUT = True
+
 
 # Local App Settings
 # ------------------------------------------------------------------------------